requirements.txt
python-telegram-bot==22.0
python-dotenv==1.0.1
requests==2.32.3
cryptography==46.0.1
